% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{import_sklearn}
\alias{import_sklearn}
\title{Import SK Learn}
\usage{
import_sklearn(assign_to_environment = T)
}
\arguments{
\item{assign_to_environment}{if \code{TRUE} assings to environment}
}
\value{
python object
}
\description{
\href{https://scikit-learn.org/stable/modules/classes.html#api-reference}{SK learn API from python}
}
