% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_calibration}
\alias{sk_calibration}
\title{SK Calibration
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.calibration}{Calibration functions}
\href{https://scikit-learn.org/stable/modules/calibration.html#calibration}{Calibration examples}}
\usage{
sk_calibration()
}
\value{
python object
}
\description{
SK Calibration
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.calibration}{Calibration functions}
\href{https://scikit-learn.org/stable/modules/calibration.html#calibration}{Calibration examples}
}
