% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_covariance}
\alias{sk_covariance}
\title{SK Covariance
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.covariance}{co-variance functions}
\href{https://scikit-learn.org/stable/modules/covariance.html#covariance}{co-variance examples}}
\usage{
sk_covariance()
}
\value{
python object
}
\description{
SK Covariance
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.covariance}{co-variance functions}
\href{https://scikit-learn.org/stable/modules/covariance.html#covariance}{co-variance examples}
}
