% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_cross_decomposition}
\alias{sk_cross_decomposition}
\title{SK Cross Decomposition
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.cross_decomposition}{Cross-decomposition functions}
\href{https://scikit-learn.org/stable/modules/cross_decomposition.html#cross-decomposition}{Cross decomposition examples}}
\usage{
sk_cross_decomposition()
}
\value{
python object
}
\description{
SK Cross Decomposition
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.cross_decomposition}{Cross-decomposition functions}
\href{https://scikit-learn.org/stable/modules/cross_decomposition.html#cross-decomposition}{Cross decomposition examples}
}
