% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_datasets}
\alias{sk_datasets}
\title{SKLearn Data}
\usage{
sk_datasets()
}
\value{
python object
}
\description{
Access to the sci-kit learn datasets
}
\details{
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.datasets}{datasets}
}
