% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_decomp}
\alias{sk_decomp}
\title{SK Decomposition
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.decomposition}{decomposition functions}
\href{https://scikit-learn.org/stable/modules/decomposition.html#decompositions}{decomposition example}}
\usage{
sk_decomp()
}
\value{
python object
}
\description{
SK Decomposition
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.decomposition}{decomposition functions}
\href{https://scikit-learn.org/stable/modules/decomposition.html#decompositions}{decomposition example}
}
