% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_discriminant_analysis}
\alias{sk_discriminant_analysis}
\title{SK Discriminant Models
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.discriminant_analysis}{Discriminant functions}
\href{https://scikit-learn.org/stable/modules/lda_qda.html#lda-qda}{Discriminant examples}}
\usage{
sk_discriminant_analysis()
}
\value{
python object
}
\description{
SK Discriminant Models
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.discriminant_analysis}{Discriminant functions}
\href{https://scikit-learn.org/stable/modules/lda_qda.html#lda-qda}{Discriminant examples}
}
