% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sk_ds_to_tibble}
\alias{sk_ds_to_tibble}
\title{Convert sci-kit learn dataset into a tibble}
\usage{
sk_ds_to_tibble(data, include_group = T, feature_names = NULL)
}
\arguments{
\item{data}{sci-kit learn data set}

\item{include_group}{if \code{TRUE} appends group index to tibble and if group is numeric strats index at 1 instead of zero}

\item{feature_names}{if not \code{NULL} tibble featre names}
}
\value{
\code{tibble()}
}
\description{
Convert sci-kit learn dataset into a tibble
}
\examples{
library(tidyverse)
library(sklearn)
ds <- sk_datasets()
n_samples <- 1500L
moons <- ds$make_moons(n_samples = n_samples, noise = .05)

df_moon <-
moons \%>\%
sk_ds_to_tibble(include_group = T, feature_names = c("x", "y", "group"))

df_moon \%>\% glimpse()
}
