% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_dummy}
\alias{sk_dummy}
\title{SK Dummy Variables
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.dummy}{Dummy variable functions}}
\usage{
sk_dummy()
}
\value{
python object
}
\description{
SK Dummy Variables
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.dummy}{Dummy variable functions}
}
