% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_ensemble}
\alias{sk_ensemble}
\title{SK Ensemble}
\usage{
sk_ensemble()
}
\value{
python object
}
\description{
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.ensemble}{ensemble functions}
\href{https://scikit-learn.org/stable/modules/ensemble.html#ensemble}{ensemble examples}
}
