% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_feature_selection}
\alias{sk_feature_selection}
\title{SK Feature Selection
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.feature_selection}{Feature selection functions}
\href{https://scikit-learn.org/stable/modules/feature_selection.html#feature-selection}{feature selection examples}}
\usage{
sk_feature_selection()
}
\value{
python object
}
\description{
SK Feature Selection
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.feature_selection}{Feature selection functions}
\href{https://scikit-learn.org/stable/modules/feature_selection.html#feature-selection}{feature selection examples}
}
