% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_gaussian_process}
\alias{sk_gaussian_process}
\title{SK Gaussian Process
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.gaussian_process}{Gaussian process functions}
\href{https://scikit-learn.org/stable/modules/gaussian_process.html#gaussian-process}{Gaussian process examples}}
\usage{
sk_gaussian_process()
}
\value{
python jobject
}
\description{
SK Gaussian Process
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.gaussian_process}{Gaussian process functions}
\href{https://scikit-learn.org/stable/modules/gaussian_process.html#gaussian-process}{Gaussian process examples}
}
