% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_impute}
\alias{sk_impute}
\title{SK Impute
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.impute}{impute functions}}
\usage{
sk_impute()
}
\value{
python object
}
\description{
SK Impute
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.impute}{impute functions}
}
