% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_isotonic}
\alias{sk_isotonic}
\title{SK Isotonic
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.isotonic}{Isotonic functions}
\href{https://scikit-learn.org/stable/modules/isotonic.html#isotonic}{Isotonic examples}}
\usage{
sk_isotonic()
}
\value{
python object
}
\description{
SK Isotonic
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.isotonic}{Isotonic functions}
\href{https://scikit-learn.org/stable/modules/isotonic.html#isotonic}{Isotonic examples}
}
