% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_kernel_approximation}
\alias{sk_kernel_approximation}
\title{SK Kernel Approxomation
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.kernel_approximation}{Kernel Approxomation functions}
\href{https://scikit-learn.org/stable/modules/kernel_approximation.html#kernel-approximation}{Kernel Approxomation examples}}
\usage{
sk_kernel_approximation()
}
\value{
python object
}
\description{
SK Kernel Approxomation
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.kernel_approximation}{Kernel Approxomation functions}
\href{https://scikit-learn.org/stable/modules/kernel_approximation.html#kernel-approximation}{Kernel Approxomation examples}
}
