% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_kernel_ridge}
\alias{sk_kernel_ridge}
\title{SK Kernel Ridge
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.kernel_ridge}{Kernel Ridge functions}
\href{https://scikit-learn.org/stable/modules/kernel_ridge.html#kernel-ridge}{Kernel Ridge examples}}
\usage{
sk_kernel_ridge()
}
\value{
python object
}
\description{
SK Kernel Ridge
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.kernel_ridge}{Kernel Ridge functions}
\href{https://scikit-learn.org/stable/modules/kernel_ridge.html#kernel-ridge}{Kernel Ridge examples}
}
