% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_lm}
\alias{sk_lm}
\title{SK Learn LM
\itemize{
\item \href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.linear_model}{linear model functions}
\item \href{https://scikit-learn.org/stable/modules/linear_model.html#linear-model}{linear model examples}
}}
\usage{
sk_lm()
}
\value{
python object
}
\description{
SK Learn LM
\itemize{
\item \href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.linear_model}{linear model functions}
\item \href{https://scikit-learn.org/stable/modules/linear_model.html#linear-model}{linear model examples}
}
}
