% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_metrics}
\alias{sk_metrics}
\title{SK Metrics
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.metrics}{metrics functions}}
\usage{
sk_metrics()
}
\value{
python object
}
\description{
SK Metrics
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.metrics}{metrics functions}
}
