% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_mixture}
\alias{sk_mixture}
\title{SK Mixture
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.mixture}{Mixture functions}}
\usage{
sk_mixture()
}
\value{
python object
}
\description{
SK Mixture
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.mixture}{Mixture functions}
}
