% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_model_selection}
\alias{sk_model_selection}
\title{SK Model Selection}
\usage{
sk_model_selection()
}
\value{
python object
}
\description{
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.covariance}{model selection functions}
\href{https://scikit-learn.org/stable/modules/covariance.html#covariance}{model selectionexamples}
}
