% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_multi_class}
\alias{sk_multi_class}
\title{SK Multi Class
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.multiclass}{multi-class functions}
\href{https://scikit-learn.org/stable/modules/multiclass.html#multiclass}{multiclass examples}}
\usage{
sk_multi_class()
}
\value{

}
\description{
SK Multi Class
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.multiclass}{multi-class functions}
\href{https://scikit-learn.org/stable/modules/multiclass.html#multiclass}{multiclass examples}
}
