% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_multioutput}
\alias{sk_multioutput}
\title{SK Multioutput
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.multioutput}{Multioutput functions}
\href{https://scikit-learn.org/stable/modules/multiclass.html#multiclass}{multioutput examples}}
\usage{
sk_multioutput()
}
\value{
python object
}
\description{
SK Multioutput
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.multioutput}{Multioutput functions}
\href{https://scikit-learn.org/stable/modules/multiclass.html#multiclass}{multioutput examples}
}
