% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_neighbors}
\alias{sk_neighbors}
\title{Neighbors}
\usage{
sk_neighbors()
}
\value{
python object
}
\description{
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.neighbors}{nearest neighbor functions}
\href{https://scikit-learn.org/stable/modules/neighbors.html#neighbors}{nearest neighbor examples}
}
