% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_nnet}
\alias{sk_nnet}
\title{SK Neural Net
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.neural_network}{Neural Network functions}
\href{https://scikit-learn.org/stable/modules/neural_networks_supervised.html#neural-networks-supervised}{Neural Network examples}}
\usage{
sk_nnet()
}
\value{
python object
}
\description{
SK Neural Net
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.neural_network}{Neural Network functions}
\href{https://scikit-learn.org/stable/modules/neural_networks_supervised.html#neural-networks-supervised}{Neural Network examples}
}
