% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_preprocessing}
\alias{sk_preprocessing}
\title{SK Learn Preprocessing}
\usage{
sk_preprocessing()
}
\value{
python object
}
\description{
Functions to pre-process data
}
\details{
\itemize{
\item \href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.preprocessing}{preprocessing functions}
\item \href{https://scikit-learn.org/stable/modules/preprocessing.html#preprocessing}{preprocessing examples}
}
}
