% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_random_projection}
\alias{sk_random_projection}
\title{SK Random Projection
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.random_projection}{Random projection functions}
\href{https://scikit-learn.org/stable/modules/random_projection.html#random-projection}{Random Projection examples}}
\usage{
sk_random_projection()
}
\value{
python object
}
\description{
SK Random Projection
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.random_projection}{Random projection functions}
\href{https://scikit-learn.org/stable/modules/random_projection.html#random-projection}{Random Projection examples}
}
