% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sk_tibble}
\alias{sk_tibble}
\title{Convert SK Matrix to Tibble}
\usage{
sk_tibble(data, feature_names = NULL)
}
\arguments{
\item{data}{SK data output}

\item{feature_names}{if not \code{NULL} data feature names}
}
\value{
\code{tibble}
}
\description{
Convert SK Matrix to Tibble
}
