% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_utils}
\alias{sk_utils}
\title{SK Utilities
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.utils}{utility functions}}
\usage{
sk_utils()
}
\value{
python object
}
\description{
SK Utilities
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.utils}{utility functions}
}
