% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn-package.r
\docType{package}
\name{sklearn}
\alias{sklearn}
\alias{sklearn-package}
\title{R port of the Python scikit-learn module}
\description{
R port of the Python scikit-learn module
}
\author{
Alex Bresler (abreslerb@asbcllc.com)
}
