% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_bayes}
\alias{sk_bayes}
\title{SK Bayes
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.naive_bayes}{Bayes functions}
\href{https://scikit-learn.org/stable/modules/naive_bayes.html#naive-bayes}{Bayes examples}}
\usage{
sk_bayes()
}
\value{
python object
}
\description{
SK Bayes
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.naive_bayes}{Bayes functions}
\href{https://scikit-learn.org/stable/modules/naive_bayes.html#naive-bayes}{Bayes examples}
}
