% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_cluster}
\alias{sk_cluster}
\title{SKLearn Clustering
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.cluster}{clustering functions}
\href{https://scikit-learn.org/stable/modules/clustering.html#clustering}{clustering examples}}
\usage{
sk_cluster()
}
\value{
python object
}
\description{
SKLearn Clustering
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.cluster}{clustering functions}
\href{https://scikit-learn.org/stable/modules/clustering.html#clustering}{clustering examples}
}
