% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sk_manifold}
\alias{sk_manifold}
\title{SK Manifold}
\usage{
sk_manifold()
}
\value{
python object
}
\description{
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.manifold}{manifold functions}
\href{https://scikit-learn.org/stable/modules/manifold.html#manifold}{manifold examples}
}
